requirements.txt
# Django and REST framework
Django>=3.2.0,<4.0.0
djangorestframework>=3.12.0,<3.14.0
djangorestframework-simplejwt>=5.0.0,<5.2.0

# Database connectors
mysqlclient==2.1.1
psycopg2-binary==2.9.5
djongo==1.3.6
pymongo==3.12.3  # Specific version for compatibility with djongo

# Additional requirements
django-filter==23.1
requests==2.28.2

# Development tools
django-debug-toolbar>=3.2.0,<3.8.0
pymysql